chore: Create a simple banking program
The python code is used to simulate a simple banking scenario.
Only the following functions are available:
    - Withdraw
    - Deposit

The user needs to enter in their details:
    - Name
    - Withdraw? -> Withdraw amount
    - Deposit? -> Deposit amount

The user is given a random UUID that is generated using python's
uuid module.
The balance is incremanted and decremented as appropriate each time a
transaction is made.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>